fix: remove inversify-express as some unresolved bug for async
