+++ tools/mock-service-host/package.json
@@ -30,9 +30,9 @@
     "convict": "^6.0.0",
     "cross-env": "^5.2.0",
     "dotenv": "^7.0.0",
+    "express": "^4.16.2",
     "fs-extra": "^7.0.1",
     "inversify": "^5.0.1",
-    "inversify-express-utils": "^6.3.2",
     "jwt-simple": "^0.5.6",
     "lodash": "^4.17.15",
     "mockjs": "^1.1.0",
