+++ tools/mock-service-host/src/router/validateAndMockRouter.ts
@@ -1,0 +1,6 @@
+import * as express from 'express'
+import { validateRequestAndMockResponse } from '../controller/validateAndMockController'
+
+export const validateAndMockRouter = express.Router()
+
+validateAndMockRouter.all('/*', validateRequestAndMockResponse)
